Renaming some components and restructuring the project.
