+++ src/app/app.routes.ts
@@ -1,10 +1,10 @@
 import { Routes } from '@angular/router';
-import { PostPageComponent } from './main-page/post-page/post-page.component';
+import { PostListingComponent } from './components/post-listing/post-listing.component';
 import { PostFormComponent } from './containers/forms/post-form/post-form.component';
-import { UserPageComponent } from './main-page/user-page/user-page.component';
+import { UserProfileComponent } from './components/user-profile/user-profile.component';
 import { SignupFormComponent } from './containers/forms/signup-form/signup-form.component';
 
-export const routes: Routes = [{path : 'post-page', component : PostPageComponent},
+export const routes: Routes = [{path : 'post-page', component : PostListingComponent},
                                {path : 'post-form', component : PostFormComponent},
                                {path : 'user-page', component : SignupFormComponent},
 ];
