+++ src/app/components/post-listing/post-listing.component.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { MainPageComponent } from './main-page.component';
+import { PostListingComponent } from './post-listing.component';
 
-describe('MainPageComponent', () => {
-  let component: MainPageComponent;
-  let fixture: ComponentFixture<MainPageComponent>;
+describe('PostPageComponent', () => {
+  let component: PostListingComponent;
+  let fixture: ComponentFixture<PostListingComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [MainPageComponent]
+      imports: [PostListingComponent]
     })
     .compileComponents();
     
-    fixture = TestBed.createComponent(MainPageComponent);
+    fixture = TestBed.createComponent(PostListingComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
