+++ src/app/containers/main/main.component.scss
